feat(tooltip): update docs and props for tooltip, definition tooltip
