+++ packages/react/src/components/Tooltip/next/DefinitionTooltip.stories.js
@@ -9,13 +9,17 @@ import './story.scss';
 
 import React from 'react';
 import { DefinitionTooltip } from './DefinitionTooltip';
+import mdx from './DefinitionTooltip.mdx';
 
 export default {
-  title: 'Components/Tooltip/DefinitionTooltip',
+  title: 'Components/DefinitionTooltip',
   component: DefinitionTooltip,
   parameters: {
     controls: {
       hideNoControlsWarning: true,
+    },
+    docs: {
+      page: mdx,
     },
     layout: 'centered',
   },
@@ -31,23 +35,29 @@ export default {
       },
     },
   },
+  decorators: [
+    (Story) => (
+      <div className="sb-tooltip-story sb-definition-tooltip">
+        <Story />
+      </div>
+    ),
+  ],
 };
 
 export const Default = () => {
   const definition =
     'Uniform Resource Locator; the address of a resource (such as a document or website) on the Internet.';
   return (
-    <div>
-      <p>
-        Custom domains direct requests for your apps in this Cloud Foundry
-        organization to a{' '}
-        <DefinitionTooltip definition={definition}>URL</DefinitionTooltip> that
-        you own. A custom domain can be a shared domain, a shared subdomain, or
-        a shared domain and host.
-      </p>
-    </div>
+    <p>
+      Custom domains direct requests for your apps in this Cloud Foundry
+      organization to a{' '}
+      <DefinitionTooltip definition={definition}>URL</DefinitionTooltip> that
+      you own. A custom domain can be a shared domain, a shared subdomain, or a
+      shared domain and host.
+    </p>
   );
 };
+
 const PlaygroundStory = (props) => {
   const { align, defaultOpen, definition } = props;
   return (
@@ -68,7 +78,7 @@ export const Playground = PlaygroundStory.bind({});
 
 Playground.argTypes = {
   align: {
-    defaultValue: 'bottom',
+    defaultValue: 'bottom-left',
     options: [
       'top',
       'top-left',
@@ -77,14 +87,6 @@ Playground.argTypes = {
       'bottom',
       'bottom-left',
       'bottom-right',
-
-      'left',
-      'left-bottom',
-      'left-top',
-
-      'right',
-      'right-bottom',
-      'right-top',
     ],
     control: {
       type: 'select',
@@ -97,5 +99,6 @@ Playground.argTypes = {
     control: {
       type: 'text',
     },
+    defaultValue: 'Example definition',
   },
 };
